fix: 불필요한 결합
Console 클래스와 VoucherService, CustomerService 클래스의 결합을 제거하였습니다. 대신 Administrator 클래스를 생성하고 ConsoleService, VoucherService, CustomerService를 결합하여 로직 수행을 위해 필요한 메서드를 호출할 수 있도록 수정하였습니다.
